BUG#13985 ndb_mgm "status" command can return incorrect data node status
Second half of the fix for this bug.

This patch forces a heartbeat to be sent and will wait (a little while)
for replies. This way we can get

> all status
X starting
Y started
X started
>

which is okay as the new status comes after the old status, always.
There is the slimmest of opportunities to get output like above where only half
the cluster appears started.

This is about the best we can do with a command line interactive program.