Bug#22384 - DELETE FROM table causes "Incorrect key file for table"
Deletes on a big index could crash the index when it needs to
shrink.

Put a forgotten negation operator in.

No test case. It is too big for the test suite. And it does not
work with 4.0, only with higher versions. It is attached to the
bug report.


myisam/mi_delete.c:
  Bug#22384 - DELETE FROM table causes "Incorrect key file for table"
  Put a negation operator ('!') before _mi_get_last_key() in del().
  It returns NULL on error, non-NULL on success.